chnage tournament controller

+++ app/controllers/api/v1/tournaments_controller.rb
@@ -4,25 +4,26 @@ module Api
       before_action :authenticate_user
       def index
         @tournaments = Tournament.all
-        render json: TournamentsRepresenter.new(@tournaments).with_teams_and_users
+        render json: TournamentsRepresenter.new(@tournaments).with_all
       end
 
       def create
         @tournament = Tournament.new(tournament_params)
         @tournament.save!
-        render json: TournamentRepresenter.new(@tournament).with_teams_and_users
+        render json: TournamentRepresenter.new(@tournament).with_all
       end
 
       def update
         authorize tournament
-        tournament.update_attributes!(tournament_params)
-        render json: TournamentRepresenter.new(tournament).with_teams_and_users
+        tournament.update_attributes!(permitted_attributes(tournament))
+        render json: TournamentRepresenter.new(tournament).with_all
       end
 
       def destroy
         if tournament.state == "ended" || tournament.state == "started"
-          render json: { errors: "Tournament started or ended"}, status: 422
+          render json: { errors: "Tournament already started or ended"}, status: 422
         else
+          authorize tournament
           tournament.destroy!
           render json: {}, status: :no_content
         end
